Use long-rtt in insert back-off policy
The backoff policy used `concurrencyLimit.getLastRtt`. This could cause
requests to be delayed way longer than necessary. For example if a
partition gets created, the last round-trip time might have been long,
but that doesn't mean that the next requests would need to get throttled
that much.

This changed the policy to use `concurrencyLimit.getLongRtt`.

This fixed a regression introduced with https://github.com/crate/crate/commit/e1aa90635a492922563cf6f8e015b69cf6e85254

cr8 run-spec specs/partitioned_bulk_insert.py localhost:4200:

Before the change:

    Runtime (in ms):
        mean:    6873.725 ± 882.141
        min/max: 747.398 → 11790.370
    Percentile:
        50:   6951.108 ± 3182.490 (stdev)
        95:   11431.199
        99.9: 11790.370

After:

    Runtime (in ms):
        mean:    755.441 ± 66.171
        min/max: 459.070 → 1502.214
    Percentile:
        50:   765.613 ± 238.725 (stdev)
        95:   1264.758
        99.9: 1502.214
